[#1447] Fix inconsistance test data deletion
[#1447] Fix invalid handler

+++ backend/api/v1/v1_data/tests/tests_data.py
@@ -97,8 +97,10 @@ class DataTestCase(TestCase):
         header = {"HTTP_AUTHORIZATION": f"Bearer {self.token}"}
 
         # NOT FOUND
-        res = self.client.delete("/api/v1/data/1", follow=True, **header)
-        self.assertEqual(res.status_code, 404)
+        exists = FormData.objects.filter(pk=1).first()
+        if not exists:
+            res = self.client.delete("/api/v1/data/1", follow=True, **header)
+            self.assertEqual(res.status_code, 404)
 
         data_id = FormData.objects.first().id
